[BE] refactor : 테스트를 위한 ObjectMapper 분리
